Prevent AOSP Temp DDS switch
Issue - When AOSP auto DDS Ui is disabled, during call
on nDDS sub, temp DDS switch is observed as a result of
which device does not have any default pdn setup while
call on nDDS is ongoing.

As device is in DSDA, data++data feature is allowed by
modem which recommends data is allowed on nDDS, due to
this recommendation, the DDS switch to nDDS sub is not
stopped.

Fix - Use AOSP logic to select preferred data sub during call
when smart DDS feature is disabled on device.

Change-Id: I7051975c3bf7de7b08e30ef47259662630031731
CRs-Fixed: 4064154