Fix Synchronization with stopScan()
There is code inside of stopScan() that attempts to
deactivate an active scan. Unfortunately, in order to
check the DB to throw IAE/ISE the scn code needs to
synchronize its data. Since stopScan() should be an
idempotent call regardless, in the event that a stop
is called on an invalid scan, we can just silently
drop them.

Bug: 284540191
Test: atest NetworkScanRequestTrackerTest
Flag: EXEMPT bugfix
Change-Id: I0510ab8347b73c1ee58ed50ca1e104be3ac23bee