Rename aconfig flag to change namespace
The change in ag/35608769 to modify the namespace of an existing flag
ended up duplicating the workflow, breaking the ability to advance the
flag.

In order for the namespace change to be correctly applied, it should
have been done with a flag rename, as discussed in b/444382568,

There is no functional change in the logic, this change is exclusively
renaming an existing flag.

The Bug field in the description makes reference to the original issue that the flag was added for, so that the new workflow doesn't get confused.

Bug: 395087895
Test: TH
Flag: com.android.internal.telephony.flags.minimal_telephony_cdm_use_board_api_level

Change-Id: Ifc8645736b8f45b289b57ab31f25e8c34e751e05