Clear satellite provision state in DB when subscriber ID changes.
When an old subscriber ID is removed from the cached provisioned list, set the satellite provision state for that subscriber ID to false in the SubscriptionManagerService database. This ensures that the database reflects the current state and is consistent with the in-memory cache.

Flag: com.android.internal.telephony.flags.fix_satellite_provision_state_out_of_sync
Bug: 444025641
Test: Add new unit test
Change-Id: I377ca16834890ebafde4db4de185ed4d0ea52679