Allow handover between IWLAN and NTN
Removed the check for Terrestrial (TN) vs. Non-Terrestrial (NTN)
networks during network evaluation.

This check existed because carriers currently do not support
handover between cellular (TN) and satellite (NTN), which
would cause the network to tear down and re-establish.

However, handover between IWLAN and NTN is possible and should
be allowed. The bug was that an IWLAN network was incorrectly
treated as a TN network, which caused it to be torn down
during a handover from Wi-Fi to satellite.

This is fixed by removing the check that evaluated a network
based on whether it was originally created on a TN or NTN.

Note in the future we will need to revisit this if carriers
start to support TN/NTN handover, especially we need to
revisit the impact on internet network.

Fixes: 452727859
Test: atest DataNetworkTest DataNetworkControllerTest
Test: Basic telephony functionality tests
Flag: EXEMPT BUGFIX
Change-Id: I72396ed6e1bea3f076ea6918ecc23014ce36c06d