Hold a Strong Ref to TelephonyCallback
This is a speculative fix that holds a strong reference to the
TelephonyCallback that is registered. The speculation is that
because the IPhoneStateListener object inside the callback only
holds a weak references to the outer callback, the outer class
is getting garbage collected; thus, when the callback fires the
mActiveDataSubId isn't getting updated. If so, this is quite
a footgun.

Bug: 452390790
Test: atest DataStallRecoveryManagerTest
Flag: com.android.internal.telephony.flags.inactive_data_network_is_not_stalled
Change-Id: Ie87548312bde68a6d3a13379e776bc69221d4437