Downgrade log from error to warning in CarrierPrivilegesTracker
CPT used to print error log when detecting a specified package can't be
found for current user. But this is not a real error because it is a
normal case when packages are installed only for other users.

The error log causes unnecessary attention and even WTF cluster error
somehow. Downgrading it to warning level here to reduce noise.

Bug: 452865760
Test: atest FrameworksTelephonyTests
Flag: EXEMPT DEBUG
Change-Id: I6396b78b33965ff2af016fde74baceacbc418de8