Harden emergency notification silence by making intent explicit
Improves the reliability of the "Do Not Show Again" action for the
"Emergency calling unavailable" notification.

The PendingIntent was using an implicit broadcast, which is not
guaranteed to be delivered on modern Android versions.

This change makes the intent explicit by setting the package,
ensuring the NotificationActionReceiver is always triggered and
the notification is silenced correctly.

Flag: EXEMPT BUGFIX
Test: manual - adb command to show emergency notif and dissmissable
Test: atest com.android.internal.telephony.CarrierServiceStateTrackerTest
Bug: 416126590
Bug: 368690328
Change-Id: Ic64ae5de3df1f1fe1f771187323d7f1a0dd7548c