Set default notification sound for incoming voicemail.
In ag/33197526 the notification channel created when migrating
from Android P did not have a default ringtone sound set.
The same code path is used for initial creation of the channel
on factory reset when the first sim is inserted, so a default
is desirable to be set.

Test: Manual test of voicemail notification channel on sim change.
Flag: EXEMPT BUGFIX
Bug: 457341758
Change-Id: I34d01ed78152478e6c920a436e3abcfb5dd3771b