Pass new Consumer<Interger> object to PointingAppController#unregisterForSatelliteTransimmionUpdates
Accept() is called more than once on Consumer resultListener passed as input to SatelliteManager#stopTransmissionUpdates. This is causing IllegalStateException.

RootCause: Consumer resultListener passed as input to SatelliteManager#stopTransmissionUpdates is being passed to both 
- PointingAppController#unregisterForSatelliteTransimmionUpdates
- CMD_STOP_SATELLITE_TRANSMISSION_UPDATES
due to which accept is called twice.

To solve this issue, a new Consumer resultListener is passed to PointingAppController#unregisterForSatelliteTransimmionUpdates

Bug: 387414816
Test: atest
FLAG: EXEMPT BUGFIX
Change-Id: I7b700b28ac44daeb6d87b94a65479c020f4c7b87