PinStorage: Support platform-managed PINs
Modify the PinStorage class to support platform-managed PINs.
There are a few differences between user-managed PINs and
platform-managed PINs:
* For platform-managed PINs, we must also store the old PIN (the one
  that was set on the SIM card prior to the platform setting its own
  PIN).
* Users need to be able to retrieve the platform-managed PIN, so they
  are stored longer than user-managed PINs (which are only stored for
  unattended reboots).
* Platform-managed PINs need to be encrypted with a long-term auth-bound
  key.
* Platform-managed PINs must be available across reboots, when the
  device was unlocked and the user initiated the reboot.

This change adds support for all the scenarios mentioned above.
It maintains the existing behaviour for the get/set methods and provides
consistent behaviour for platform-managed PINs (with the additional
restriction that platform-managed PINs can only be retrieved when the
device is unlocked).

Bug: 381850338
Bug: 346544096
Test: atest FrameworksTelephonyTests:com.android.internal.telephony.uicc.PinStorageTest
Flag: android.security.auto_sim_pin_management
Change-Id: If30579b40b341bd2e1df34a1a44d248619d16b25