Assign slot index of remote SIMs to correct constant in addSubInfo
SLOT_INDEX_FOR_REMOTE_SIM_SUB exists as its own constant, but is
not the slot index assigned for remote SIMs in addSubInfo. Instead,
INVALID_SIM_SLOT_INDEX is assigned. This is only ok at the moment
since the two constants are equal. This will not be ok when they
separate to different values.

There is precedent for the use of SLOT_INDEX_FOR_REMOTE_SIM_SUB
(getSubId).

Bug: 462821889
Flag: EXEMPT PURE_REFACTOR
Test: atest SubscriptionManagerServiceTest
Test: atest MapClientContentTest
Test: atest MapClientServiceTest
Test: atest SubscriptionManagerTest
Change-Id: Idccf4b3c1d769bba4fea6ba1862739b5ef395dd5