Remove `-printusage` from telephony proguard
This is far too verbose, and is unnecessary given that the usage output
is implicitly generated in `.intermediates`. Also remove the dormant
`-printconfiguration` comment in the proguard file; the configuration is
also implicitly generated from the target's intermediates.

Note: This was discovered as part of early integration of R8 proguard
validation rules into the build system.

Test: m
Bug: 452120003
Flag: EXEMPT HOST_MODE
Change-Id: Iaa6555f6984a1c072cb95678c5fab7a16a807163