Add implementation for notifying domain selection emergency mode
This change is to add an implementation for a new TelephonyCallback API,
DomainSelectionEmergencyModeListener, which allows listeners to monitor
the emergency mode state when AP domain selection is enabled.
EmergencyStateTracker will deliver entrance and exit events to the
application corresponding to the start and end of an emergency call or
SMS respectively.

Bug: 423761083
Flag: com.android.internal.telephony.flags.domain_selection_emergency_mode_notification
Test: atest FrameworksTelephonyTests
Test: manual (verify the entrance/exit events when an emergency call/SMS
is being started and ended)

Change-Id: I9bf1d858276b95e159bd457483cfcc39fb978f0d