Fix race condition by moving data call list indication
This change moves the data call list change indication to the response interface.
Previously, this indication could arrive before the setup data call response, leading to a race condition. This update ensures a deterministic ordering of events.

This also adds a mechanism to distinguish between legacy and updated behavior for data call disconnections.
When the new behavior is active, the system marks the data call as inactive instead of immediately removing it, waiting for a subsequent indication to finalize the removal.
Directly removing the data call from the list without changing the status to inactive would not result in data call disconnected.

Bug: 438335285
Test: atest CtsTelephonyTestCases
Flag: com.android.internal.telephony.flags.support_explicit_data_disconnect
Change-Id: Ic42aac9bf62b5b15701d73d3a53dd80145ebd660