Refactor default SMS/MMS app fetching to use MAIN user
The default SMS/MMS app fetching logic is corrected to explicitly use
the MAIN user instead of implicitly relying on the default SYSTEM user.

This is necessary because Messaging support is predominantly limited to
the MAIN user (with exceptions for Enterprise work profile telephony).
The code is refactored to use MAIN user when fetching the default SMS
and MMS applications.

Detailed RCA: go/hsum-messages-failing-rca
Bug: 452616003
Test: Manually verified incoming SMS/MMS are received on a HSUM device
Flag: EXEMPT BUGFIX
Change-Id: Ib1df5958d539a2387ed5c2762bdd13e87ef48655