Support multiple remote SIMs in addSubInfo and removeSubInfo
In addSubInfo and removeSubInfo, replace the usage of
mSlotIndexToSubId with the new SubscriptionSet variable,
mRemoteSubIds. Account for this in the unit test helper
insertSubscription.

This CL is part of the topic for fixing multiple remote SIMs:
multiple_remote_sims

Bug: 455587208
Flag: com.android.internal.telephony.flags.remote_sim_sub_id_set
Test: atest SubscriptionManagerServiceTest
Test: atest MapClientContentTest
Test: atest MapClientServiceTest
Test: atest SubscriptionManagerTest

Change-Id: Ia0fd7318e82b1e6364926bdd5f983bde87dbd8d5