Prioritize local SIMs in updateDefaultSubId
Currently, if the initial sub ID in updateDefaultSubId is not
active, the first one from getActiveSubIdList is chosen. Without
guarantees on ordering, it is not certain whether a local or
remote SIM will be the first in the list.

updateDefaultSubId will be changed to prioritize local SIMs before
remote SIMs. This is done by creating getActiveLocalSubIdList and
getActiveRemoteSubIdList, which stream sub IDs from
mSlotIndexToSubId and mRemoteSubIds, respectively.

The stream of sub IDs is passed to a new method
(filterSubIdStreamVisibilityAsUser) which extracts the original
logic, irrespective of stream, into a helper.

This CL is part of the topic for fixing multiple remote SIMs:
multiple_remote_sims

Bug: 455587208
Flag: com.android.internal.telephony.flags.remote_sim_sub_id_set
Test: atest SubscriptionManagerServiceTest
Test: atest MapClientContentTest
Test: atest MapClientServiceTest
Test: atest SubscriptionManagerTest
Change-Id: I4c68146ee6ceb4771889be8699ae75717a7afdaa